Update placeholder and validation messages, improve language support
The commit introduces several adjustments to the input fields in multiple pages, such as updating the Invitation Code's placeholder and changing the button positions in login.php. Enhanced multilanguage support has been added for validation messages and labels in form fields, enabling the application to provide feedback in both English and Slovak. Also, the equality comparisons in the if conditions are corrected to triple equals sign (===) for strict comparison.