Update redirection and add cookie consent dialogue
Changed the redirection URL from login page to the homepage in "redirectToLogin" function in restricted.php. Also, updated the associated redirection messages. Implemented a new script in index.php to handle cookie consent dialogue prompting users to accept or decline the use of cookies on their devices. If declined, the user is redirected to the restricted page.