Update placeholder text and content translations
The placeholder text for the invitation code input field was updated from 'XXXXXX' to a more instructive phrase. Additionally, translations for the 'invitationHeading', 'invitationMessage', and 'whatsGoodHeading' lines have been updated and improved. Emojis have been added for a friendly touch in the 'invitationHeading' and 'whatsGoodHeading' lines for both languages, English and Slovak.