Upgrade password validation and improve UX
In this update, the password validation process was enhanced. We've added frontend password functionality to check if the new password and confirm password inputs are valid and match and then accordingly enables or disables the save changes button. Labels have been given ID's for better identification and manipulation, and messages were introduced to guide the user through the password rules and potential mismatch between new password and confirmation field. We also increased accessibility by adding IDs to buttons and provided multilingual support for the new labels and messages.